proto: avoid duplicate SigInput creation
